+++ src/duk_api_codec.c
@@ -2,18 +2,94 @@
  *  Encoding and decoding basic formats: hex, base64.
  *
  *  These are in-place operations which may allow an optimized implementation.
+ *
+ *  Base-64: https://tools.ietf.org/html/rfc4648#section-4
  */
 
 #include "duk_internal.h"
 
-/* dst length must be exactly ceil(len/3)*4 */
-DUK_LOCAL void duk__base64_encode_helper(const duk_uint8_t *src, const duk_uint8_t *src_end,
-                                         duk_uint8_t *dst, duk_uint8_t *dst_end) {
+#if defined(DUK_USE_BASE64_FASTPATH)
+DUK_LOCAL const duk_uint8_t duk__base64_enc_lookup[64] = {
+	0x41, 0x42, 0x43, 0x44, 0x45, 0x46, 0x47, 0x48, 0x49, 0x4a, 0x4b, 0x4c, 0x4d, 0x4e, 0x4f, 0x50,  /* A...P */
+	0x51, 0x52, 0x53, 0x54, 0x55, 0x56, 0x57, 0x58, 0x59, 0x5a, 0x61, 0x62, 0x63, 0x64, 0x65, 0x66,  /* Q...f */
+	0x67, 0x68, 0x69, 0x6a, 0x6b, 0x6c, 0x6d, 0x6e, 0x6f, 0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76,  /* g...v */
+	0x77, 0x78, 0x79, 0x7a, 0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37, 0x38, 0x39, 0x2b, 0x2f   /* w.../ */
+};
+DUK_LOCAL const duk_int8_t duk__base64_dec_lookup[256] = {
+	/* -1 = error, -2 = allowed whitespace, -3 = padding ('='), 0...63 decoded bytes */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -2, -2, -1, -1, -2, -1, -1,  /* 0x00...0x0f */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0x10...0x1f */
+	-2, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 62, -1, -1, -1, 63,  /* 0x20...0x2f */
+	52, 53, 54, 55, 56, 57, 58, 59, 60, 61, -1, -1, -1, -3, -1, -1,  /* 0x30...0x3f */
+	-1,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 14,  /* 0x40...0x4f */
+	15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, -1, -1, -1, -1, -1,  /* 0x50...0x5f */
+	-1, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40,  /* 0x60...0x6f */
+	41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, -1, -1, -1, -1, -1,  /* 0x70...0x7f */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0x80...0x8f */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0x90...0x9f */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0xa0...0xaf */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0xb0...0xbf */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0xc0...0xcf */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0xd0...0xdf */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,  /* 0xe0...0xef */
+	-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1   /* 0xf0...0xff */
+};
+#endif  /* DUK_USE_BASE64_FASTPATH */
+
+#if defined(DUK_USE_BASE64_FASTPATH)
+DUK_LOCAL void duk__base64_encode_helper(const duk_uint8_t *src, duk_size_t srclen, duk_uint8_t *dst) {
+	duk_uint_fast32_t t;
+	duk_size_t n_full, n_final;
+
+	n_full = srclen / 3;  /* full 3-byte -> 4-char conversions */
+	n_final = srclen - n_full * 3;
+	DUK_ASSERT_DISABLE(n_final >= 0);
+	DUK_ASSERT(n_final <= 2);
+
+	while (n_full > 0) {
+		n_full--;
+
+		t = (duk_uint_fast32_t) (*src++);
+		t = (t << 8) + (duk_uint_fast32_t) (*src++);
+		t = (t << 8) + (duk_uint_fast32_t) (*src++);
+
+		*dst++ = duk__base64_enc_lookup[t >> 18];
+		*dst++ = duk__base64_enc_lookup[(t >> 12) & 0x3f];
+		*dst++ = duk__base64_enc_lookup[(t >> 6) & 0x3f];
+		*dst++ = duk__base64_enc_lookup[t & 0x3f];
+	}
+
+	switch (n_final) {
+	/* case 0: nop */
+	case 1: {
+		/* XX== */
+		t = (duk_uint_fast32_t) (*src++);
+		*dst++ = duk__base64_enc_lookup[t >> 2];           /* XXXXXX-- */
+		*dst++ = duk__base64_enc_lookup[(t << 4) & 0x3f];  /* ------XX */
+		*dst++ = DUK_ASC_EQUALS;
+		*dst++ = DUK_ASC_EQUALS;
+		break;
+	}
+	case 2: {
+		/* XXX= */
+		t = (duk_uint_fast32_t) (*src++);
+		t = (t << 8) + (duk_uint_fast32_t) (*src++);
+		*dst++ = duk__base64_enc_lookup[t >> 10];          /* XXXXXX-- -------- */
+		*dst++ = duk__base64_enc_lookup[(t >> 4) & 0x3f];  /* ------XX XXXX---- */
+		*dst++ = duk__base64_enc_lookup[(t << 2) & 0x3f];  /* -------- ----XXXX */
+		*dst++ = DUK_ASC_EQUALS;
+		break;
+	}
+	}
+}
+#else  /* DUK_USE_BASE64_FASTPATH */
+DUK_LOCAL void duk__base64_encode_helper(const duk_uint8_t *src, duk_size_t srclen, duk_uint8_t *dst) {
 	duk_small_uint_t i, snip;
 	duk_uint_fast32_t t;
 	duk_uint_fast8_t x, y;
+	const duk_uint8_t *src_end;
 
-	DUK_UNREF(dst_end);
+	src_end = src + srclen;
 
 	while (src < src_end) {
 		/* read 3 bytes into 't', padded by zero */
@@ -58,22 +134,159 @@ DUK_LOCAL void duk__base64_encode_helper(const duk_uint8_t *src, const duk_uint8
 				y = '/';
 			}
 
-			DUK_ASSERT(dst < dst_end);
 			*dst++ = (duk_uint8_t) y;
 		}
 	}
 }
+#endif  /* DUK_USE_BASE64_FASTPATH */
 
-DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, const duk_uint8_t *src_end,
-                                               duk_uint8_t *dst, duk_uint8_t *dst_end, duk_uint8_t **out_dst_final) {
+#if defined(DUK_USE_BASE64_FASTPATH)
+DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, duk_size_t srclen, duk_uint8_t *dst, duk_uint8_t **out_dst_final) {
+	duk_int_fast32_t x;
+	duk_uint_fast32_t t;
+	duk_small_uint_t n_equal;
+	duk_small_uint_t n_chars;
+	const duk_uint8_t *src_end;
+	const duk_uint8_t *src_end_safe;
+
+	src_end = src + srclen;
+	src_end_safe = src_end - 4;  /* if 'src < src_end_safe', safe to read 4 bytes */
+
+	/* Innermost fast path processes 4 valid base-64 characters at a time
+	 * but bails out on whitespace, padding chars ('=') and invalid chars.
+	 * Once the slow path segment has been processed, we return to the
+	 * inner fast path again.  This handles e.g. base64 with newlines
+	 * reasonably well because the majority of a line is in the fast path.
+	 */
+	for (;;) {
+		/* Fast path, handle units with just actual encoding characters. */
+
+		while (src <= src_end_safe) {
+			/* The lookup byte is intentionally sign extended to (at least)
+			 * 32 bits and then ORed unsigned.  This ensures that is at least
+			 * 1 byte is negative, the highest bit of 't' will be set at the
+			 * end and we don't need to check every byte.
+			 */
+			DUK_DDD(DUK_DDDPRINT("fast loop: src=%p, src_end_safe=%p, src_end=%p",
+			                     (const void *) src, (const void *) src_end_safe, (const void *) src_end));
+
+			t = (duk_uint_fast32_t) (duk_int_fast32_t) duk__base64_dec_lookup[*src++];
+			t = (t << 6) | (duk_uint_fast32_t) (duk_int_fast32_t) duk__base64_dec_lookup[*src++];
+			t = (t << 6) | (duk_uint_fast32_t) (duk_int_fast32_t) duk__base64_dec_lookup[*src++];
+			t = (t << 6) | (duk_uint_fast32_t) (duk_int_fast32_t) duk__base64_dec_lookup[*src++];
+
+			if (DUK_UNLIKELY(t & 0x80000000UL)) {
+				DUK_DDD(DUK_DDDPRINT("fast loop unit was not clean, process one slow path unit"));
+				src -= 4;
+				break;
+			}
+
+			DUK_ASSERT(t <= 0xffffffUL);
+			DUK_ASSERT((t >> 24) == 0);
+			*dst++ = (duk_uint8_t) (t >> 16);
+			*dst++ = (duk_uint8_t) ((t >> 8) & 0xff);
+			*dst++ = (duk_uint8_t) (t & 0xff);
+		}
+
+		/* Handle one slow path unit (or finish if we're done). */
+
+		n_equal = 0;
+		n_chars = 0;
+		t = 0;
+		for (;;) {
+			DUK_DDD(DUK_DDDPRINT("slow loop: src=%p, src_end=%p, n_chars=%ld, n_equal=%ld, t=%ld",
+			                     (const void *) src, (const void *) src_end, (long) n_chars, (long) n_equal, (long) t));
+
+			if (DUK_UNLIKELY(src >= src_end)) {
+				goto done;  /* two level break */
+			}
+
+			x = duk__base64_dec_lookup[*src++];
+			if (DUK_UNLIKELY(x < 0)) {
+				if (x == -2) {
+					continue;  /* allowed ascii whitespace */
+				} else if (x == -3) {
+					n_equal++;
+					t <<= 6;
+				} else {
+					DUK_ASSERT(x == -1);
+					goto error;
+				}
+			} else {
+				DUK_ASSERT(x >= 0 && x <= 63);
+				if (n_equal > 0) {
+					/* Don't allow actual chars after equal sign. */
+					goto error;
+				}
+				t = (t << 6) + x;
+			}
+
+			if (DUK_UNLIKELY(n_chars == 3)) {
+				/* Emit 3 bytes and backtrack if there was padding.  There's
+				 * always space for the whole 3 bytes so no check needed.
+				 */
+				DUK_ASSERT(t <= 0xffffffUL);
+				DUK_ASSERT((t >> 24) == 0);
+				*dst++ = (duk_uint8_t) (t >> 16);
+				*dst++ = (duk_uint8_t) ((t >> 8) & 0xff);
+				*dst++ = (duk_uint8_t) (t & 0xff);
+
+				if (DUK_UNLIKELY(n_equal > 0)) {
+					DUK_ASSERT(n_equal <= 4);
+
+					/* There may be whitespace between the equal signs. */
+					if (n_equal == 1) {
+						/* XXX= */
+						dst -= 1;
+					} else if (n_equal == 2) {
+						/* XX== */
+						dst -= 2;
+					} else {
+						goto error;  /* invalid padding */
+					}
+
+					/* Continue parsing after padding, allows concatenated,
+					 * padded base64.
+					 */
+				}
+				break;  /* back to fast loop */
+			} else {
+				n_chars++;
+			}
+		}
+	}
+ done:
+	DUK_DDD(DUK_DDDPRINT("done; src=%p, src_end=%p, n_chars=%ld",
+	                     (const void *) src, (const void *) src_end, (long) n_chars));
+
+	DUK_ASSERT(src == src_end);
+
+	if (n_chars != 0) {
+		/* Here we'd have the option of decoding unpadded base64
+		 * (e.g. "xxxxyy" instead of "xxxxyy==".  Currently not
+		 * accepted.
+		 */
+		goto error;
+	}
+
+	*out_dst_final = dst;
+	return 1;
+
+ error:
+	return 0;
+}
+#else  /* DUK_USE_BASE64_FASTPATH */
+DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, duk_size_t srclen, duk_uint8_t *dst, duk_uint8_t **out_dst_final) {
 	duk_uint_fast32_t t;
 	duk_uint_fast8_t x, y;
 	duk_small_uint_t group_idx;
+	duk_small_uint_t n_equal;
+	const duk_uint8_t *src_end;
 
-	DUK_UNREF(dst_end);
-
+	src_end = src + srclen;
 	t = 0;
 	group_idx = 0;
+	n_equal = 0;
 
 	while (src < src_end) {
 		x = *src++;
@@ -89,42 +302,14 @@ DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, const duk
 		} else if (x == '/') {
 			y = 63;
 		} else if (x == '=') {
-			/* We don't check the zero padding bytes here right now.
-			 * This seems to be common behavior for base-64 decoders.
+			/* We don't check the zero padding bytes here right now
+			 * (that they're actually zero).  This seems to be common
+			 * behavior for base-64 decoders.
 			 */
 
-			if (group_idx == 2) {
-				/* xx== -> 1 byte, t contains 12 bits, 4 on right are zero */
-				t = t >> 4;
-				DUK_ASSERT(dst < dst_end);
-				*dst++ = (duk_uint8_t) t;
-
-				if (src >= src_end) {
-					goto error;
-				}
-				x = *src++;
-				if (x != '=') {
-					goto error;
-				}
-			} else if (group_idx == 3) {
-				/* xxx= -> 2 bytes, t contains 18 bits, 2 on right are zero */
-				t = t >> 2;
-				DUK_ASSERT(dst < dst_end);
-				*dst++ = (duk_uint8_t) ((t >> 8) & 0xff);
-				DUK_ASSERT(dst < dst_end);
-				*dst++ = (duk_uint8_t) (t & 0xff);
-			} else {
-				goto error;
-			}
-
-			/* Here we can choose either to end parsing and ignore
-			 * whatever follows, or to continue parsing in case
-			 * multiple (possibly padded) base64 strings have been
-			 * concatenated.  Currently, keep on parsing.
-			 */
-			t = 0;
-			group_idx = 0;
-			continue;
+			n_equal++;
+			t <<= 6;  /* shift in zeroes */
+			goto skip_add;
 		} else if (x == 0x09 || x == 0x0a || x == 0x0d || x == 0x20) {
 			/* allow basic ASCII whitespace */
 			continue;
@@ -132,16 +317,38 @@ DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, const duk
 			goto error;
 		}
 
+		if (n_equal > 0) {
+			/* Don't allow mixed padding and actual chars. */
+			goto error;
+		}
 		t = (t << 6) + y;
+	 skip_add:
 
 		if (group_idx == 3) {
 			/* output 3 bytes from 't' */
-			DUK_ASSERT(dst < dst_end);
 			*dst++ = (duk_uint8_t) ((t >> 16) & 0xff);
-			DUK_ASSERT(dst < dst_end);
 			*dst++ = (duk_uint8_t) ((t >> 8) & 0xff);
-			DUK_ASSERT(dst < dst_end);
 			*dst++ = (duk_uint8_t) (t & 0xff);
+
+			if (DUK_UNLIKELY(n_equal > 0)) {
+				/* Backtrack. */
+				DUK_ASSERT(n_equal <= 4);
+				if (n_equal == 1) {
+					dst -= 1;
+				} else if (n_equal == 2) {
+					dst -= 2;
+				} else {
+					goto error;  /* invalid padding */
+				}
+
+				/* Here we can choose either to end parsing and ignore
+				 * whatever follows, or to continue parsing in case
+				 * multiple (possibly padded) base64 strings have been
+				 * concatenated.  Currently, keep on parsing.
+				 */
+				n_equal = 0;
+			}
+
 			t = 0;
 			group_idx = 0;
 		} else {
@@ -163,6 +370,7 @@ DUK_LOCAL duk_bool_t duk__base64_decode_helper(const duk_uint8_t *src, const duk
  error:
 	return 0;
 }
+#endif  /* DUK_USE_BASE64_FASTPATH */
 
 /* Shared handling for encode/decode argument.  Fast path handling for
  * buffer and string values because they're the most common.  In particular,
@@ -206,8 +414,7 @@ DUK_EXTERNAL const char *duk_base64_encode(duk_context *ctx, duk_idx_t index) {
 	dstlen = (srclen + 2) / 3 * 4;
 	dst = (duk_uint8_t *) duk_push_fixed_buffer(ctx, dstlen);
 
-	duk__base64_encode_helper((const duk_uint8_t *) src, (const duk_uint8_t *) (src + srclen),
-	                          dst, (dst + dstlen));
+	duk__base64_encode_helper((const duk_uint8_t *) src, srclen, dst);
 
 	ret = duk_to_string(ctx, -1);
 	duk_replace(ctx, index);
@@ -244,12 +451,11 @@ DUK_EXTERNAL void duk_base64_decode(duk_context *ctx, duk_idx_t index) {
 	if (srclen > 4294967292UL) {
 		goto type_error;
 	}
-	dstlen = (srclen + 3) / 4 * 3;  /* upper limit */
+	dstlen = (srclen + 3) / 4 * 3;  /* upper limit, assuming no whitespace etc */
 	dst = (duk_uint8_t *) duk_push_dynamic_buffer(ctx, dstlen);
 	/* Note: for dstlen=0, dst may be NULL */
 
-	retval = duk__base64_decode_helper((const duk_uint8_t *) src, (const duk_uint8_t *) (src + srclen),
-	                                   dst, dst + dstlen, &dst_final);
+	retval = duk__base64_decode_helper((const duk_uint8_t *) src, srclen, dst, &dst_final);
 	if (!retval) {
 		goto type_error;
 	}